+++ mypy.ini
@@ -59,9 +59,6 @@ ignore_errors = True
 [mypy-lib,lib.*]
 ignore_missing_imports = True
 
-[mypy-linter_lib.printer]
-ignore_missing_imports = True
-
 [mypy-zulint,zulint.*,zulint.lister.*]
 ignore_missing_imports = True
